refactor(middlewares): improve admin and join check middleware logic
- Update import paths to use new aiogram BaseMiddleware location
- Add proper unwrapping of decorated handlers
- Implement better admin access control with module and command checks
- Add error handling for unauthorized access attempts
- Update user model to use BigInteger for IDs
- Add error logging and rollback for database operations

+++ app/db/models/user.py
@@ -1,11 +1,10 @@
-from sqlalchemy import Column, Integer, String, Boolean
+from sqlalchemy import Column, BigInteger, String, Boolean
 from app.db.base import Base
 
 class User(Base):
     __tablename__ = "users"
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     username = Column(String, nullable=True)
     is_admin = Column(Boolean, default=False)
     is_banned = Column(Boolean, default=False)
-    referred_by = Column(Integer, nullable=True)
-
+    referred_by = Column(BigInteger, nullable=True)
